@@ -2,7 +2,7 @@
 name = "assignments"
 version = "0.0.1"
 readme = "README.md"
-requires-python = ">=3.8"
+requires-python = ">=3.10"
 description = "First assignment of the FAAST Advance Foundations"
 authors = [
     {name = "Fernando Cordeiro<fernando@daredata.engineering>"}
